+++ src/components/Home/About/index.tsx
@@ -4,6 +4,7 @@ import { useTranslations } from "next-intl";
 
 import Terminal from "../../Common/Terminal";
 import TimeLines from "../../Common/TimeLines";
+import HeaderArticle from "@/components/Common/HeaderArticle";
 
 interface IAboutProps {}
 
@@ -29,20 +30,10 @@ const About: FC<IAboutProps> = ({}) => {
     <div className="overflow-hidden">
       <section
         id="about"
-        className="mx-auto px-4 py-4 lg:container lg:px-20 lg:py-20"
+        className="mx-auto px-4 pb-4 lg:container lg:px-20 lg:pb-20"
       >
-        <div className="container mx-auto">
-          <div className="-mx-4 flex flex-wrap">
-            <div className="w-full px-4">
-              <div className="mx-auto mb-4 text-center">
-                <h2 className="mb-4 text-3xl font-bold text-primaryText-900 sm:text-4xl md:text-[40px]">
-                  {t("title")}
-                </h2>
-              </div>
-            </div>
-          </div>
-        </div>
-        <article className="grid grid-cols-1 gap-4 lg:grid-cols-2 lg:gap-16 lg:py-8">
+        <HeaderArticle title={t("title")} description="" subtitle="" />
+        <article className="grid grid-cols-1 gap-4 lg:grid-cols-2 lg:gap-16 lg:pb-8">
           <div className="space-y-4">
             <p className="text-primaryText text-center text-base">
               {t("info.description1")}